requirements.txt
# Core graph + RAG pipeline
# Core NLP / RAG / KG pipeline
PyMuPDF
groq
langchain
langchain-community
langchain-text-splitters
sentence-transformers
scikit-learn
networkx
numpy
chromadb

# Graph community detection
python-igraph
leidenalg

# Visualization
pyvis

# Image classification
torch
torchvision
pillow

# Web App
streamlit

